split GA component
cohesion of functions
