+++ src/pages/_app.tsx
@@ -3,6 +3,7 @@ import type { AppProps } from 'next/app';
 import { useEffect, useState } from 'react';
 import { ScoreMapProvider } from 'components/ScoreMapContext';
 import Script from 'next/script';
+import GA from 'components/GA';
 
 function MyApp({ Component, pageProps }: AppProps) {
   const [height, setHeight] = useState('100vh');
@@ -20,15 +21,7 @@ function MyApp({ Component, pageProps }: AppProps) {
       className="fullscreen-container"
       style={{ minHeight: height, height: 'auto' }}
     >
-      <Script
-        src={`https://www.googletagmanager.com/gtag/js?id=${process.env.NEXT_PUBLIC_GOOGLE_ANALYTICS_ID}`}
-      />
-      <Script id="ga-setting">{`
-  window.dataLayer = window.dataLayer || [];
-  function gtag(){dataLayer.push(arguments);}
-  gtag('js', new Date());
-  gtag('config', '${process.env.NEXT_PUBLIC_GOOGLE_ANALYTICS_ID}');
-      `}</Script>
+      <GA />
       <div className="container" style={{ height: height }}>
         <ScoreMapProvider>
           <Component {...pageProps} />
